requirements.txt
h5py
matplotlib
numpy
opencv_python
Pillow
Pillow
PyYAML
PyYAML
scipy
timm
torch
torch_optimizer
torchvision
tqdm
